move enum schema to main file
